+++ src/lib.rs
@@ -2,6 +2,12 @@
 #![feature(type_alias_impl_trait)]
 #![feature(min_specialization)]
 
+#[cfg(feature = "std")]
+extern crate std;
+
+#[cfg(not(feature = "std"))]
+extern crate core as std;
+
 pub mod environment;
 pub mod font;
 pub mod layout;
